+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - ReleaseDate
 ### Fixed
 - network: unrouted requests don't lead to infinite waiting anymore ([#116]).
+- network: release routed if flow closed.
 
 [#116]: https://github.com/elfo-rs/elfo/pull/116